+++ KD_Lib/KD/common/__init__.py
