+++ KD_Lib/KD/common/base_class.py
