+++ KD_Lib/KD/vision/TAKD/__init__.py
