+++ KD_Lib/KD/vision/noisy/noisy_teacher.py
@@ -1,13 +1,19 @@
+import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from KD_Lib.RKD import RKDLoss
-from KD_Lib.common import BaseClass
+
+import random
+from copy import deepcopy
+import matplotlib.pyplot as plt
+
+from .utils import add_noise
+from KD_Lib.KD.common import BaseClass
 
 
-class VanillaKD(BaseClass):
+class NoisyTeacher(BaseClass):
     """
-    Original implementation of Knowledge distillation from the paper "Distilling the
-    Knowledge in a Neural Network" https://arxiv.org/pdf/1503.02531.pdf
+    Implementation of Knowledge distillation using a noisy teacher from the paper "Deep
+    Model Compression: Distilling Knowledge from Noisy Teachers" https://arxiv.org/pdf/1610.09650.pdf
 
     :param teacher_model (torch.nn.Module): Teacher model
     :param student_model (torch.nn.Module): Student model
@@ -15,6 +21,8 @@ class VanillaKD(BaseClass):
     :param val_loader (torch.utils.data.DataLoader): Dataloader for validation/testing
     :param optimizer_teacher (torch.optim.*): Optimizer used for training teacher
     :param optimizer_student (torch.optim.*): Optimizer used for training student
+    :param alpha (float): Threshold for deciding if noise needs to be added
+    :param noise_variance (float): Variance parameter for adding noise
     :param loss_fn (torch.nn.Module):  Calculates loss during distillation
     :param temp (float): Temperature parameter for distillation
     :param distil_weight (float): Weight paramter for distillation loss
@@ -31,6 +39,8 @@ class VanillaKD(BaseClass):
         val_loader,
         optimizer_teacher,
         optimizer_student,
+        alpha=0.5,
+        noise_variance=0.1,
         loss_fn=nn.MSELoss(),
         temp=20.0,
         distil_weight=0.5,
@@ -38,7 +48,7 @@ class VanillaKD(BaseClass):
         log=False,
         logdir="./Experiments",
     ):
-        super(VanillaKD, self).__init__(
+        super(NoisyTeacher, self).__init__(
             teacher_model,
             student_model,
             train_loader,
@@ -53,6 +63,9 @@ class VanillaKD(BaseClass):
             logdir,
         )
 
+        self.alpha = alpha
+        self.noise_variance = noise_variance
+
     def calculate_kd_loss(self, y_pred_student, y_pred_teacher, y_true):
         """
         Function used for calculating the KD loss during distillation
@@ -62,12 +75,12 @@ class VanillaKD(BaseClass):
         :param y_true (torch.FloatTensor): Original label
         """
 
-        soft_teacher_out = F.softmax(y_pred_teacher / self.temp, dim=1)
-        soft_student_out = F.softmax(y_pred_student / self.temp, dim=1)
+        if random.uniform(0, 1) <= self.alpha:
+            y_pred_teacher = add_noise(y_pred_teacher, self.noise_variance)
 
-        loss = (1 - self.distil_weight) * F.cross_entropy(y_pred_student, y_true)
+        loss = (1.0 - self.distil_weight) * F.cross_entropy(y_pred_student, y_true)
         loss += (self.distil_weight * self.temp * self.temp) * self.loss_fn(
-            soft_teacher_out, soft_student_out
+            F.log_softmax(y_pred_student / self.temp, dim=1),
+            F.softmax(y_pred_teacher / self.temp, dim=1),
         )
-
         return loss
