+++ KD_Lib/KD/vision/vanilla/__init__.py
@@ -1,0 +1,1 @@
+from .vanilla_kd import VanillaKD
