--- KD_Lib/attention/__init__.py
@@ -1,2 +1,0 @@
-from .loss_metric import ATLoss
-from .attention import attention
