--- KD_Lib/vanilla/__init__.py
@@ -1,1 +1,0 @@
-from KD_Lib.vanilla.vanilla_kd import VanillaKD
