+++ README.rst
@@ -5,7 +5,7 @@ KD_Lib
 .. image:: https://travis-ci.com/SforAiDl/KD_Lib.svg?branch=master
     :target: https://travis-ci.com/SforAiDl/KD_Lib
 
-A Pytorch Library to help extend all Knowledge Distillation works
+A PyTorch library to easily facilitate knowledge distillation for custom deep learning models
 
 Installation :
 ==============
@@ -43,34 +43,34 @@ Currently implemented works
 +-----------------------------------------------------------+----------------------------------+----------------------+
 |  Paper                                                    |  Link                            | Repository (KD_Lib/) |
 +===========================================================+==================================+======================+
-| Distilling the Knowledge in a Neural Network              | https://arxiv.org/abs/1503.02531 | vanilla              |
+| Distilling the Knowledge in a Neural Network              | https://arxiv.org/abs/1503.02531 | KD/vision/vanilla    |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Improved Knowledge Distillation via Teacher Assistant     | https://arxiv.org/abs/1902.03393 | TAKD                 |
+| Improved Knowledge Distillation via Teacher Assistant     | https://arxiv.org/abs/1902.03393 | KD/vision/TAKD       |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Relational Knowledge Distillation                         | https://arxiv.org/abs/1904.05068 | RKD                  |
+| Relational Knowledge Distillation                         | https://arxiv.org/abs/1904.05068 | KD/vision/RKD        |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Distilling Knowledge from Noisy Teachers                  | https://arxiv.org/abs/1610.09650 | noisy                |
+| Distilling Knowledge from Noisy Teachers                  | https://arxiv.org/abs/1610.09650 | KD/vision/noisy      |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Paying More Attention To The Attention                    | https://arxiv.org/abs/1612.03928 | attention            |
+| Paying More Attention To The Attention                    | https://arxiv.org/abs/1612.03928 | KD/vision/attention  |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Revisit Knowledge Distillation: a Teacher-free Framework  | https://arxiv.org/abs/1909.11723 | teacher_free         |
+| Revisit Knowledge Distillation: a Teacher-free Framework  | https://arxiv.org/abs/1909.11723 |KD/vision/teacher_free|
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Mean Teachers are Better Role Models                      | https://arxiv.org/abs/1703.01780 | mean_teacher         |
+| Mean Teachers are Better Role Models                      | https://arxiv.org/abs/1703.01780 |KD/vision/mean_teacher|
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Knowledge Distillation via Route Constrained Optimization | https://arxiv.org/abs/1904.09149 | RCO                  |
+| Knowledge Distillation via Route Constrained Optimization | https://arxiv.org/abs/1904.09149 | KD/vision/RCO        |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Born Again Neural Networks                                | https://arxiv.org/abs/1805.04770 | BANN                 |
+| Born Again Neural Networks                                | https://arxiv.org/abs/1805.04770 | KD/vision/BANN       |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Preparing Lessons: Improve Knowledge Distillation with    | https://arxiv.org/abs/1911.07471 | KA                   |
+| Preparing Lessons: Improve Knowledge Distillation with    | https://arxiv.org/abs/1911.07471 | KD/vision/KA         |
 | Better Supervision                                        |                                  |                      |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Improving Generalization Robustness with Noisy            | https://arxiv.org/abs/1910.05057 | noisy                |
+| Improving Generalization Robustness with Noisy            | https://arxiv.org/abs/1910.05057 | KD/vision/noisy      |
 | Collaboration in Knowledge Distillation                   |                                  |                      |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Distilling Task-Specific Knowledge from BERT into         | https://arxiv.org/abs/1903.12136 | Bert2Lstm            |
+| Distilling Task-Specific Knowledge from BERT into         | https://arxiv.org/abs/1903.12136 | KD/text/BERT2LSTM    |
 | Simple Neural Networks                                    |                                  |                      |
 +-----------------------------------------------------------+----------------------------------+----------------------+
-| Deep Mutual Learning                                      | https://arxiv.org/abs/1706.00384 | DML                  |
+| Deep Mutual Learning                                      | https://arxiv.org/abs/1706.00384 | KD/vision/DML        |
 +-----------------------------------------------------------+----------------------------------+----------------------+
 | The Lottery Ticket Hypothesis: Finding                    | https://arxiv.org/abs/1803.03635 | Pruning/             |
 | Sparse, Trainable Neural Networks                         |                                  | lottery_tickets      |
