+++ logs.rst
@@ -40,7 +40,7 @@ Results for Knowledge Distillation Algorithms -
 |                        |                                  |  ResNet18 - Student          |             213178             |                      |     0.9641          |
 +------------------------+----------------------------------+------------------------------+--------------------------------+----------------------+---------------------+
 |  SelfTraining          |               MNIST              |  ResNet18 - Student          |             213178             |         5            |     0.9772          |
-+------------------------+----------------------------------+-------------------------------+--------------------------------+----------------------+--------------------+
++------------------------+----------------------------------+------------------------------+--------------------------------+----------------------+---------------------+
 |  VirtualTeacher        |               MNIST              |  ResNet18 - Student          |             213178             |         5            |     0.9814          |
 +------------------------+----------------------------------+------------------------------+--------------------------------+----------------------+---------------------+
 |  RCO                   |               MNIST              |  ResNet50 - Teacher          |             454010             |                      |     0.9827          |
